+++ src/main/java/subway/repository/query/SectionQuery.java
@@ -1,4 +1,4 @@
-package subway.query;
+package subway.repository.query;
 
 public class SectionQuery {
     public static final String insertQuery = "insert into section (line_id, up_station_id, down_station_id, distance) values(?, ?, ?, ?)";
