+++ src/main/java/subway/service/LineService.java
@@ -1,7 +1,7 @@
 package subway.service;
 
 import org.springframework.stereotype.Service;
-import subway.dao.LineDao;
+import subway.repository.dao.LineDao;
 import subway.domain.Line;
 import subway.exception.DuplicateException;
 import subway.exception.NotFoundException;
