+++ src/test/java/subway/station/StationDaoTest.java
@@ -7,7 +7,7 @@ import org.springframework.boot.test.autoconfigure.jdbc.AutoConfigureTestDatabas
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.annotation.DirtiesContext;
 import org.springframework.test.context.junit.jupiter.SpringExtension;
-import subway.dao.StationDao;
+import subway.repository.dao.StationDao;
 import subway.domain.Station;
 
 import java.util.Optional;
